pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-tumor-classifier-render"
version = "1.0.0"
description = "Flask inference server for Brain Tumor MRI classification"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
  "flask>=3.0.0",
  "flask-cors>=4.0.0",
  "gunicorn>=21.2.0",
  "numpy>=1.24.0",
  "Pillow>=10.0.0",
  "torch==2.11.0",
  "torchvision==0.26.0",
  "efficientnet-pytorch>=0.7.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0"
]

[tool.setuptools]
py-modules = ["app", "config"]
